==> 1_emissions.ipynb <==
import pandas as pd
import numpy as np
import pickle

# Read 2023 emisison data
scope1_2023=pd.read_csv('data/CDP/scope1_2023.csv').set_index('Account number')
scope2_2023=pd.read_csv('data/CDP/scope2_2023.csv').set_index('Account number')
scope3_2023=pd.read_csv('data/CDP/scope3_2023.csv')

# Read 2018 emisison data
scope1_2018=pd.read_csv('data/CDP/scope1_2018.csv').set_index('Account number')
scope2_2018=pd.read_csv('data/CDP/scope2_2018.csv').set_index('Account number')

# Read companies and extract all IDs
target_car=pd.read_excel('data/company_data/Data.xlsx', sheet_name='Targets_car')
target_og=pd.read_excel('data/company_data/Data.xlsx', sheet_name='Targets_OG')
target_airline=pd.read_excel('data/company_data/Data.xlsx', sheet_name='Targets_Airlines')

company_list=list(target_car['CDP_ID'].unique())+list(target_og['CDP_ID'].unique())+list(target_airline['CDP_ID'].unique())
company_list= [x for x in company_list if str(x) != 'nan']

# Read years for which companies report
reporting_year=pd.read_csv('data/CDP/reporting_year.csv')

# Read data that was manually added due to missing CDP records
missing_2017=pd.read_excel('data/company_data/Data.xlsx', sheet_name='Missing_2017').set_index('CDP_ID')

#missing_2018=pd.read_excel('data/company_data/Data.xlsx', sheet_name='Missing_2018').set_index('CDP_ID')


## Scope 1 Emissions

#Compile dataframe for all scope 1 emissions (i.e. 2023 and 2018)

scope1_total=pd.DataFrame(index=company_list, columns=['company_name', 'scope1_2023', 'scope1_2018'])
for i in scope1_total.index:
    if i in scope1_2023.index:
        print(i)
        
        scope1_total.loc[i, 'company_name']=scope1_2023.loc[i, 'Organization']
        scope1_total.loc[i, 'scope1_2023']=scope1_2023.loc[i, 'scope1_emissions']
       
      

    if i in scope1_2018.index:
        
        scope1_total.loc[i, 'scope1_2018']=scope1_2018.loc[i, 'scope1_emissions']
        


for i in missing_2017.index:
    scope1_total.loc[i, 'scope1_2018']=missing_2017.loc[i, 'scope1']
   # scope1_total.loc[i, 'scope1_2018_assurance']=missing_2018.loc[i, 'scope1_assurance']


scope1_total['scope1_2023']=pd.to_numeric(scope1_total['scope1_2023'], errors='coerce')
scope1_total['scope1_2018']=pd.to_numeric(scope1_total['scope1_2018'], errors='coerce')

scope1_total['scope1_difference']=scope1_total['scope1_2023']/scope1_total['scope1_2018']



#save to pickle
scope1_total.to_pickle('data/temp/scope1_total.pkl')

scope1_total[scope1_total['scope1_2018'].isna()]

scope1_total.loc[6855]

## Scope 2 emissions


#compile dataframe for scop2 emissions including 2018 and 2023 data and add missing values
scope2_total=pd.DataFrame(index=company_list, columns=['company_name', 'scope2_2023',  'scope2_2018'])

for i in scope2_total.index:
    if i in scope2_2023.index:

        
        scope2_total.loc[i, 'company_name']=scope2_2023.loc[i, 'Organization']
        scope2_total.loc[i, 'scope2_2023']=scope2_2023.loc[i, 'scope2_location_based']
       
        




        

    if i in scope2_2018.index:
        
        scope2_total.loc[i, 'scope2_2018']=scope2_2018.loc[i, 'scope2_location_based']

        if scope2_total.loc[i, 'scope2_2018'] == 'Question not applicable':
            scope2_total.loc[i, 'scope2_2018']=scope2_2018.loc[i, 'scope2_market_based']
       



for i in missing_2017.index:
        if missing_2017.loc[i, 'scope2_location_based'] is not np.nan:
            scope2_total.loc[i, 'scope2_2018']=missing_2017.loc[i, 'scope2_location_based']
        else:
            scope2_total.loc[i, 'scope2_2018']=missing_2017.loc[i, 'scope2_undefined']


        
  

scope2_total['scope2_2023']=pd.to_numeric(scope2_total['scope2_2023'], errors='coerce')
scope2_total['scope2_2018']=pd.to_numeric(scope2_total['scope2_2018'], errors='coerce')

scope2_total['scope2_difference']=scope2_total['scope2_2023']/scope2_total['scope2_2018']




scope2_total.to_pickle('data/temp/scope2_total.pkl')

len(scope2_total[scope2_total['scope2_2018'].isna()])
