==> 2_scope3_shares.ipynb <==
import pandas as pd

# Read targets
car=pd.read_excel('data/company_data/Data.xlsx',sheet_name='Targets_car')
OG=pd.read_excel('data/company_data/Data.xlsx',sheet_name='Targets_OG')
Airlines=pd.read_excel('data/company_data/Data.xlsx',sheet_name='Targets_Airlines')

#Read CDP data on emissions
scope3_emission=pd.read_csv('data/CDP/scope3_2023.csv')
scope1_emissions=pd.read_csv('data/CDP/scope1_2023.csv')
scope2_emissions=pd.read_csv('data/CDP/scope2_2023.csv')

#only keep most recent records
scope1_emissions=scope1_emissions[scope1_emissions['RowName']=='Reporting year']
scope2_emissions=scope2_emissions[scope2_emissions['RowName']=='Reporting year']

#make a list of relevant companies
relevant_car=list(car['CDP_ID'].unique())
relevant_OG=list(OG['CDP_ID'].unique())
relevant_Airlines=list(Airlines['CDP_ID'].unique())

#filter for relevant companies
scope3_car=scope3_emission[scope3_emission['Account number'].isin(relevant_car)]
scope3_OG=scope3_emission[scope3_emission['Account number'].isin(relevant_OG)]
scope3_Airlines=scope3_emission[scope3_emission['Account number'].isin(relevant_Airlines)]

scope1_car=scope1_emissions[scope1_emissions['Account number'].isin(relevant_car)]
scope1_OG=scope1_emissions[scope1_emissions['Account number'].isin(relevant_OG)]
scope1_Airlines=scope1_emissions[scope1_emissions['Account number'].isin(relevant_Airlines)]

scope2_car=scope2_emissions[scope2_emissions['Account number'].isin(relevant_car)]
scope2_OG=scope2_emissions[scope2_emissions['Account number'].isin(relevant_OG)]
scope2_Airlines=scope2_emissions[scope2_emissions['Account number'].isin(relevant_Airlines)]

scope3_car=scope3_car[~scope3_car['RowName'].str.contains('Other')]
scope3_og=scope3_OG[~scope3_OG['RowName'].str.contains('Other')]
scope3_Airlines=scope3_Airlines[~scope3_Airlines['RowName'].str.contains('Other')]

## Calculate average scope 3 shares

# make dataframe
df_scope3_by_cat=pd.DataFrame(index=scope3_car['RowName'].unique())

import pandas as pd

def calculate_average_share(scope3_emission):
    # Convert 'scope3_emissions' column to numeric if it's not already numeric
    scope3_emission['scope3_emissions'] = pd.to_numeric(scope3_emission['scope3_emissions'], errors='coerce')
    filtered_df = scope3_emission[~scope3_emission['evaluation_status'].isin(['Not evaluated', 'Relevant, not yet calculated'])]

    # Group data by company
    grouped = filtered_df.groupby('Organization')

    # Initialize an empty DataFrame to store results
    results = pd.DataFrame()

    # Iterate over each group (company)
    for company, group_data in grouped:
        # Calculate total Scope 3 emissions for the company
        total_scope3_emissions = group_data['scope3_emissions'].sum()

        # Calculate the share of each category in total Scope 3 emissions
        group_data['category_share'] = group_data['scope3_emissions'] / total_scope3_emissions

        # Append the results to the overall results DataFrame
        results = pd.concat([results, group_data])

    # Calculate the average share of each category across all companies
    average_share_by_category = results.groupby('RowName')['category_share'].mean()
    
    # Normalize shares to sum up to 1
    total_shares = average_share_by_category.sum()
    average_share_by_category = average_share_by_category / total_shares


    return average_share_by_category

#Calculate average share of scope 3 by category per sectors
df_scope3_by_cat['car']=calculate_average_share(scope3_car)
df_scope3_by_cat['OG']=calculate_average_share(scope3_OG)
df_scope3_by_cat['Airlines']=calculate_average_share(scope3_Airlines)


df_scope3_by_cat.to_pickle('data/temp/scope3_by_cat.pkl')



df_scope3_by_cat

## Calculate share of scopes of total emissions

# Aggrecate scope 3 emissions by company
def get_tot_scope3(df):
    df=df.reset_index()

    df_return=pd.DataFrame(columns=['scope3_emissions'], index=df['Account number'].unique())

    
    df['scope3_emissions']=pd.to_numeric(df['scope3_emissions'], errors='coerce') 

    df_return['scope3_emissions']=df.groupby('Account number')['scope3_emissions'].sum()

    return df_return


# get share of total emissions by scope
def get_shares(scope1, scope2, scope3):
    scope1['scope1_emissions']=pd.to_numeric(scope1['scope1_emissions'], errors='coerce')
    scope2['scope2_location_based']=pd.to_numeric(scope2['scope2_location_based'], errors='coerce')

    exclude_list=scope1[scope1['scope1_emissions'].isna()].index
    exclude_list=exclude_list.append(scope2[scope2['scope2_location_based'].isna()].index)
    exclude_list=exclude_list.append(scope3[scope3['scope3_emissions']==0].index)

    scope1=scope1[~scope1.index.isin(exclude_list)]
    scope2=scope2[~scope2.index.isin(exclude_list)]
    scope3=scope3[~scope3.index.isin(exclude_list)]


    total_emissions = scope1['scope1_emissions'].sum()+scope2['scope2_location_based'].sum()+scope3['scope3_emissions'].sum()
    share_scope1=scope1['scope1_emissions'].sum()/total_emissions
    share_scope2=scope2['scope2_location_based'].sum()/total_emissions
    share_scope3=scope3['scope3_emissions'].sum()/total_emissions

    return [share_scope1, share_scope2, share_scope3]








#calculate share of scopes
share_car=get_shares(scope1_car, scope2_car, get_tot_scope3(scope3_car))
share_OG=get_shares(scope1_OG, scope2_OG, get_tot_scope3(scope3_OG))
share_Airline = get_shares(scope1_Airlines, scope2_Airlines, get_tot_scope3(scope3_Airlines))

df_scope_share=pd.DataFrame([share_car, share_OG, share_Airline], columns=['Scope 1', 'Scope 2', 'Scope 3'], index=['Automobile', 'O&G', 'Airline'])

df_scope_share.to_pickle('data/temp/scope_share.pkl')
scope3_emission.to_pickle('data/temp/scope3.pkl')

df_scope_share

df_scope_share

(df_scope3_by_cat*100)
