==> 4_control_variables.ipynb <==
import pandas as pd

#read control variables
df_og_capital=pd.read_excel('data/company_data/control_og.xlsx', sheet_name='Sheet1')
df_car_capital=pd.read_excel('data/company_data/control_car.xlsx', sheet_name='Sheet1')
df_airline_capital=pd.read_excel('data/company_data/control_airline.xlsx', sheet_name='Sheet1')

#read company data
df_og_cdp=pd.read_excel('data/company_data/Data.xlsx', sheet_name='O&G')
df_car_cdp=pd.read_excel('data/company_data/Data.xlsx', sheet_name='Car')
df_airline_cdp=pd.read_excel('data/company_data/Data.xlsx', sheet_name='Airlines')

import pandas as pd

# change columan names
def transform_dataframe(df):
    # Extract the first row which contains the years
    years = df.iloc[0].values

    # Define the function to process the column names
    def process_column_name(col_name, year):
        col_name = col_name.replace("SP_", "").replace("IQ_", "")
        col_name = ''.join([i for i in col_name if not i.isdigit() and i != '.'])
        return f"{col_name}_{year}"

    # Apply the function to all columns except the first two
    new_columns = [df.columns[0], df.columns[1]] + [
        process_column_name(col, year) for col, year in zip(df.columns[2:], years[2:])
    ]

    # Update the dataframe with the new column names
    df.columns = new_columns

    # Drop the first row as it's no longer needed
    df = df.drop(index=0)

    return df

# transform dataframes
df_og_capital = transform_dataframe(df_og_capital)
df_car_capital = transform_dataframe(df_car_capital)
df_airline_capital = transform_dataframe(df_airline_capital)

#merge our company data with s&p catpiel iq pro data
df_og_capital=pd.merge(df_og_capital, df_og_cdp, on='SP_ENTITY_ID', how='inner')
df_car_capital=pd.merge(df_car_capital, df_car_cdp, on='SP_ENTITY_ID', how='inner')
df_airline_capital=pd.merge(df_airline_capital, df_airline_cdp, on='SP_ENTITY_ID', how='inner')


#combine sectors
df_combined = pd.concat([df_og_capital, df_car_capital, df_airline_capital], ignore_index=True)


len(df_airline_capital.columns)

len(df_og_capital.columns)

for column in df_og_capital.columns:
    if column not in df_airline_capital.columns:
        print(column)

df_og_capital

df_airline_capital.columns

#save
df_og_capital.to_pickle('data/temp/og_control.pkl')
df_car_capital.to_pickle('data/temp/car_control.pkl')
df_airline_capital.to_pickle('data/temp/airline_control.pkl')

df_combined.to_pickle('data/temp/combined_control.pkl')

df_og_capital
